BUG: be explicit about the types for numerical calculations
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10128
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10128 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/Wizard/vtkEMSegmentRunSegmentationStep.cxx
@@ -940,7 +940,7 @@ void vtkEMSegmentRunSegmentationStep::ProcessRunRegistrationOutputGUIEvents(
 
       for(int i=0; i<numPts; i++,ptr++)
       {
-        if(*ptr) { *ptr = log(*ptr); }
+        if(*ptr) { *ptr = static_cast<unsigned int> (log(static_cast<float>(*ptr))); }
       }
 
       double range[2];
